feat: Add open transition to dropdown menu
- Modify Modal to be a simple dynamic Portal
- Move original modal styles to base dropdown
