+++ client/components/Dropdown/Base/BaseDropdown.scss
@@ -1,3 +1,17 @@
+.dropdown {
+  position: absolute;
+  width: 100%;
+  max-height: 0;
+  overflow-y: hidden;
+  background-color: white;
+  box-shadow: 0 0 4px 1px rgba(252, 120, 120, 0.01), 0 3px 24px rgba(0, 0, 0, 0.2);
+  transition: max-height 0.2s ease;
+
+  &--open {
+    max-height: 500px;
+  }
+}
+
 .header {
   padding: 16px 24px;
   border-bottom: 1px solid #eee;
@@ -26,6 +40,7 @@
 
   &--selected {
     color: #1591b6;
+    background-color: #fafafa;
 
     &::after {
       position: absolute;
