--- client/components/Modal/Modal.scss
@@ -1,6 +1,0 @@
-.modal {
-  position: absolute;
-  background-color: white;
-  box-shadow: 0 0 4px 1px rgba(252, 120, 120, 0.01), 0 3px 24px rgba(0, 0, 0, 0.2);
-  overflow: hidden;
-}
