+++ client/shared/dropdown.ts
@@ -1,0 +1,14 @@
+interface Style {
+  [propName: string]: string;
+}
+
+export interface Menu {
+  text: string;
+  value: string;
+  style?: Style;
+}
+
+export interface Selected {
+  data: string;
+  time: string;
+}
