+++ client/shared/styleProps.ts
@@ -1,0 +1,11 @@
+export interface Position {
+  top?: number | string;
+  right?: number | string;
+  bottom?: number | string;
+  left?: number | string;
+}
+
+export interface Box {
+  width?: number | string;
+  height?: number | string;
+}
