+++ public/index.html
@@ -11,7 +11,7 @@
   </head>
   <body>
     <div id="app"></div>
-    <div id="modal"></div>
+    <div id="dropdown"></div>
 
     <!-- Main -->
     <script src="bundle.js"></script>
